+++ src/index.css
@@ -1,2 +1,2 @@
-@import './about.css';
-@import './app.css';
+@import './about/about.css';
+@import './home/home.css';
