+++ src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import './index.css';
 
-import App from './app';
-import About from './about';
+import Home from './home/home';
+import About from './about/about';
 import reportWebVitals from './reportWebVitals';
 import { Link } from "react-router-dom";
 
@@ -16,7 +16,7 @@ ReactDOM.render(
         <Link to="/about">About</Link>
       </nav>
       <Routes>
-        <Route path="/" element={<App />} />
+        <Route path="/" element={<Home />} />
         <Route path="about" element={<About />} />
       </Routes>
     </BrowserRouter>
